backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
google-cloud-speech==2.21.0
numpy==1.24.3
sounddevice==0.4.6
scipy==1.11.3
webrtcvad==2.0.10
pydub==0.25.1
python-multipart==0.0.6
websockets==11.0.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9